update func handle insert timsheets

+++ erpnext/projects/doctype/timesheet/timesheet.py
@@ -4,6 +4,7 @@
 import datetime
 import json
 import asyncio
+import copy
 
 import frappe
 from frappe import _
@@ -584,19 +585,20 @@ async def handler_insert_timesheets():
             time_sheet_doc.company = "ACONS"
             time_sheet_doc.employee = emp_name
             time_sheet_doc.status = TIME_SHEET_STATUS[task_status]
+            total_hours = copy.deepcopy(time_sheet_doc.total_hours)
 
             if len(dates) > 0:
                 prev_time_logs = time_sheet_doc.time_logs
-                for row in prev_time_logs:
-                    if convert_date_to_datetime(row.from_time) not in dates:
-                        print("total hours", time_sheet_doc.total_hours)
-                        print("prev hours", row.hours)
-                        time_sheet_doc.total_hours -= flt(row.hours)
-                        frappe.db.delete("Timesheet Detail", row.name)
+
+                if len(prev_time_logs) > 0:
+                    for row in prev_time_logs:
+                        if convert_date_to_datetime(row.from_time) not in dates:
+                            total_hours -= flt(row.hours)
+                            frappe.db.delete("Timesheet Detail", row.name)
 
                 for date, hrs in dates.items():
                     curr_log = next((row for row in prev_time_logs if convert_date_to_datetime(row.from_time) == date), None)
-                    time_sheet_doc.total_hours += flt(hrs)
+                    total_hours += flt(hrs)
 
                     if curr_log is None:
                         time_sheet_doc.append(
@@ -618,6 +620,7 @@ async def handler_insert_timesheets():
                     curr_log.task = task_doc.name
                     curr_log.completed = task_status == "Completed"
 
+            time_sheet_doc.total_hours = total_hours
             time_sheet_doc.insert() if is_new_time_sheet else time_sheet_doc.save()              
             if task_status == "Completed": time_sheet_doc.submit()
             excel_data_update[row_num] = f"{new_hash_key}--{time_sheet_doc.name}"
